requirements.txt
django
gunicorn
django-heroku
django_q
django-ckeditor
django_countries
django_crispy_forms
django_redis
django-dotenv
boto3